+++ griffon/output.py
@@ -329,9 +329,6 @@ def text_output_products_contain_component(
     if "results" in output and output["count"] > 0:
         console.highlighter = None
 
-        # order at the end
-        # ordered_results = sorted(output["results"], key=lambda d: d["product_stream"])
-
         # first flatten the tree
         normalised_results = generate_normalised_results(
             output,
@@ -349,11 +346,13 @@ def text_output_products_contain_component(
             console.highlighter = None
 
             flaw_mode = ctx.params["flaw_mode"]
-            flaw_operation = "dry_run"
+            flaw_operation = "new"
             if flaw_mode == "add":
                 flaw_operation = "new"
             if flaw_mode == "update":
                 flaw_operation = "update"
+            if flaw_mode == "dry_run":
+                flaw_operation = "dry_run"
 
             generate_affects(ctx, result_tree, exclude_components, flaw_operation, no_wrap=False)