+++ griffon/commands/queries.py
@@ -272,6 +272,13 @@ def retrieve_component_summary(ctx, component_name, strict_name_search):
     help="Do not search middleware.",
 )
 @click.option(
+    "--no-upstream-affects",
+    "no_upstream_affects",
+    is_flag=True,
+    default=False,
+    help="Do not generate upstream affects.",
+)
+@click.option(
     "--include-inactive-product-streams",
     "include_inactive_product_streams",
     is_flag=True,
@@ -321,6 +328,7 @@ def get_product_contain_component(
     search_upstreams,
     no_community,
     no_middleware,
+    no_upstream_affects,
     include_inactive_product_streams,
     include_product_stream_excluded_components,
     output_type_filter,