+++ CHANGELOG.md
@@ -5,9 +5,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+
+## [0.2.20] - 2023-08-11
 ### Changed
 * make 'new' default for -a (when generating affects)
-
 ### Added
 * --no-upstream-affects which will exclude upstream components when generating affects